[Dotnet Monitor] Small Fix For SDM Triggers (#4332)
Small bug fix on the dotnet-monitor side from the unification work - we
previously didn't create a payload for `BeginInstrumentReporting`, and
thus only checked that a payload wasn't an error or the end of a
counter. The code now leverages the newly-added `IsValuePublishedEvent`
from the unification to only apply to the correct payloads.