Allow ` in hex strings (#4379)
Update extension command parsing to allow a single \` in the middle of
addresses. We allow "\`0123" because the user might have copied the \`
marker in "0000\`0123". This makes commands more consistent since the
C++ based SOS commands allow for \` in the addresses.

Updated a location where we didn't use TryParseAddress but should have.

Feature request from the GC team.