Move the symstore repo to diagnostics (#4603)
This PR moves the main symstore assemblies, tests and dotnet-symbol tool
to the diagnostics repo. It doesn't move the NugetSymbolServer or the
SymClient test tool. Neither are necessary for symbol indexing and
downloading.

The docs/specs were not copied either. I really don't want two copies of
the indexing and download protocol specs.

There is a base line commit that is exactly what is currently in the
symstore repro and several other commits that fix all the coding
convention and analyzer changes required to live in the diagnostics
repo. The only real functional change in
dotnet-symbol/Microsoft.SymbolStore is that ConfigureAwait(false) was
added to all the awaits (I think the default without ConfigureAwait is
true).