Provide info about PDB container type (MSF vs. MSFZ) (#4877)
This is a follow-on to my previous PR, #4868 . This allows users of the
`PDBFile` class to query for the "kind" of the container that a given
PDB file uses. Normal (uncompressed) PDBs use a container format called
"MSF". Compressed PDBs use a container format called "MSFZ".

This adds two properties to `PDBFile`. The `ContainerKind` property
returns an enum which allows the caller to distinguish between MSF, and
MSFZ. This also allows for new container formats (as yet unspecified) in
the future. However, the `ContainerKind` enum intentionally does not
distinguish between the version of the container. For that, the
`ContainerKindSpecString` returns a string which contains a textual
identifier ("msf" vs. "msfz") and for MSFZ, a version number. For now,
the only version number defined is 0.

---------

Co-authored-by: Arlie Davis <ardavis@microsoft.com>