Add support for creating a DiagnosticClient from a diagnostic port (#5070)
Fixes #3631

Add support for creating a `DiagnosticClient` from a diagnostic port
string. Currently we have to rely on `System.Reflection`, so it makes it
hacky today to create our own advanced tooling on top of
`DiagnosticClient`, while all the tooling developed in this repo are
able to access internal APIs.