Implement SaveAllModules (#5069)
Fixes #3138

This is my tentative at implementing a `!SaveAllModules` command to save
all managed modules from a debugging target.

The coding conventions are all over the place in strike.cpp, I tried my
best to fit in.
I moved the logic of the `!SaveModule` command to a `SaveModuleToFile`
helper, so I could reuse it for the new command. `SaveModulesFromDomain`
enumerates the modules in an appdomain and calls `SaveModuleToFile` for
each. The `!SaveAllModules` command then just calls
`SaveModulesFromDomain` for each appdomain.