[dotnet-trace][collect-linux] Reduce status update frequency (#5620)
The RecordTrace FFI callback can be invoked on the magnitude of several
thousands of times every second. Updating the progress status is too
aggressive, and caused all of the events to be dropped. Instead,
updating the progress status every second 1. actually leads to a
different elapsed time and 2. allows adequate event throughput.